feat(code-source): add initial  implementation.  Bind code source to workspace and create envs branches if they do not exist

+++ internal/common/easytool/util/id_util/id_util.go
@@ -3,6 +3,6 @@ package id_util
 var snowflake = NewSnowflake(1)
 
 // GenID use snowflake to generate global unique id
-func GenID() int64 {
+func GenID() int {
 	return snowflake.GenerateID()
 }
